+++ apis/nucleus/src/components/listbox/ListBox.jsx
@@ -129,6 +129,7 @@ export default function ListBox({
     selections,
     checkboxes,
     doc: document,
+    loaderRef,
   });
 
   const onFocus = (event) =>